# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VitalGuide Health AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  name: string;
  age: number;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  medicalConditions?: string | null;
  /** @nullable */
  medications?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  activityLevel?: string | null;
  /** @nullable */
  goals?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt?: string;
}

export interface UserProfileInput {
  name: string;
  age: number;
  gender?: string;
  weight?: number;
  height?: number;
  medicalConditions?: string;
  medications?: string;
  allergies?: string;
  sleepHours?: number;
  activityLevel?: string;
  goals?: string;
  location?: string;
}

export interface Plan {
  id: number;
  title: string;
  type: string;
  /** @nullable */
  description?: string | null;
  status: string;
  createdAt: string;
  updatedAt?: string;
}

export interface PlanInput {
  title: string;
  type: string;
  description?: string;
}

export interface PlanUpdate {
  title?: string;
  status?: string;
  description?: string;
}

export interface DailyLog {
  id: number;
  /** @nullable */
  planId?: number | null;
  logDate: string;
  /** @nullable */
  mood?: string | null;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  waterIntake?: number | null;
  /** @nullable */
  foodLog?: string | null;
  /** @nullable */
  symptomsLog?: string | null;
  /** @nullable */
  medicationTaken?: boolean | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface DailyLogInput {
  planId?: number;
  logDate?: string;
  mood?: string;
  sleepHours?: number;
  waterIntake?: number;
  foodLog?: string;
  symptomsLog?: string;
  medicationTaken?: boolean;
  notes?: string;
}

export interface Conversation {
  id: number;
  mode: string;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  mode: string;
  title: string;
}

export interface Message {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessage {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface SendOpenaiMessageInput {
  content: string;
}