# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VitalGuide Health AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: Get daily health logs (last 7 days)
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
    post:
      operationId: createLog
      tags: [logs]
      summary: Create a daily health log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyLogInput"
      responses:
        "201":
          description: Log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
  /logs/today:
    get:
      operationId: getTodayLog
      tags: [logs]
      summary: Get today's log entry
      responses:
        "200":
          description: Today's log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
        "404":
          description: No log for today
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List AI conversations
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [chat]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get AI response (streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response stream (text/event-stream)
...
components:
  schemas:
...
    DailyLog:
      type: object
      required: [id, logDate, createdAt]
      properties:
        id:
          type: integer
        planId:
          type: ["integer", "null"]
        logDate:
          type: string
        mood:
          type: ["string", "null"]
        sleepHours:
          type: ["number", "null"]
        waterIntake:
          type: ["number", "null"]
        foodLog:
          type: ["string", "null"]
        symptomsLog:
          type: ["string", "null"]
        medicationTaken:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DailyLogInput:
      type: object
      properties:
        planId:
          type: integer
        logDate:
          type: string
        mood:
          type: string
        sleepHours:
          type: number
        waterIntake:
          type: number
        foodLog:
          type: string
        symptomsLog:
          type: string
        medicationTaken:
          type: boolean
        notes:
          type: string
    Conversation:
      type: object
      required: [id, mode, title, createdAt]
      properties:
        id:
          type: integer
        mode:
          type: string
        title:
          type: string
        createdAt:
          type: string
    ConversationInput:
      type: object
      required: [mode, title]
      properties:
        mode:
          type: string
        title:
          type: string
    Message:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...